rename main2.py to main.py
